backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pydantic-settings==2.10.1
python-dotenv==1.1.1
numpy==2.3.2
scipy==1.16.2
packaging==25.0
networkx==3.5
scikit-fuzzy==0.5.0
